requirements.txt
appdirs>=1.4.2
asn1crypto>=0.24.0
attrs>=17.4.0
blinker>=1.4
certifi>=2018.8.24
cffi>=1.11.5
chardet>=3.0.4
click>=6.7
cryptography>=2.3
cryptography-vectors>=2.3.1
enum34>=1.1.6
Flask>=1.0.2
Flask-SQLAlchemy>=2.2
Flask-SSO>=0.4.0
Flask-WTF>=0.14.2
funcsigs>=1.0.2
honcho>=0.7.1
idna>=2.7
ipaddress>=1.0.22
itsdangerous>=0.24
Jinja2>=2.10
MarkupSafe>=0.23
MySQL-python>=1.2.5
packaging>=16.8
pluggy>=0.6.0
py>=1.5.2
pycparser>=2.18
PyJWT>=1.6.4
PyMySQL>=0.7.10
pyparsing>=2.1.10
pytest>=3.4.0
requests>=2.18.1
six>=1.11.0
SQLAlchemy>=1.1.6
urllib3>=1.21.1
uWSGI>=2.0.14
Werkzeug>=0.14.1
WTForms>=2.1
